fix: make event loop handle send errors correctly
When the crossterm event stream is dropped, or the application event
channel is dropped, the event loop now correctly exits.

If the init event fails to send, the application will likely not work
correctly, so the event loop now panics if the init event fails to send.

Refactored the event loop to make it easier to read and understand.